Move concat implementation to separate file, propagate generic types to concat result DF
